+++ src/vite-env.d.ts
@@ -1,0 +1,13 @@
+/// <reference types="vite/client" />
+
+interface Window {
+  lmEditor: {
+    backend: {
+      baseUrl: string
+    }
+    versions: {
+      electron: string
+      node: string
+    }
+  }
+}